Make ClrThreadPool and hill climbing compile on both Windows and Unix as per PR comments.
